renaming local and rest query
