+++ lib/providers/bridge.js
@@ -1,5 +1,5 @@
-const RestQuery = require('./query/restquery');
-const LocalQuery = require('./query/localquery');
+const RestQuery = require('./query/rest-query');
+const LocalQuery = require('./query/local-query');
 const RestFetchRequest = require('./fetch/rest-fetch-request');
 const LocalFetchRequest = require('./fetch/local-fetch-request');
 const ModuleDigest = require('../digest/module-digest.js');
