+++ lib/providers/query/local-query.js
