+++ lib/providers/query/rest-query.js
